+++ plugins/workloads/pkg/info.go
@@ -10,7 +10,6 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/client-go/dynamic"
 	"k8s.io/client-go/kubernetes"
 	_ "k8s.io/client-go/plugin/pkg/client/auth"
@@ -129,7 +128,12 @@ func CreateResourceProviderFromAPI(ctx context.Context, dynamicClient dynamic.In
 		return nil, err
 	}
 
-	workloads, err := controller.GetAllTopControllers(ctx, dynamicClient, restMapper, "")
+	client := controller.Client{
+		Context: ctx,
+		Dynamic: dynamicClient,
+		RESTMapper: restMapper,
+	}
+	workloads, err := client.GetAllTopControllersSummary("")
 	if err != nil {
 		logrus.Errorf("Error while getting all TopControllers: %v", err)
 		return nil, err
@@ -138,22 +142,15 @@ func CreateResourceProviderFromAPI(ctx context.Context, dynamicClient dynamic.In
 	for _, workload := range workloads {
 		topController := workload.TopController
 		var containers []ContainerResult
-		podCount := float64(len(workload.Pods))
-		pod := workload.Pods[0]
-		// Convert the unstructured object to cluster.
-		var pd corev1.Pod
-		err = runtime.DefaultUnstructuredConverter.
-			FromUnstructured(pod.UnstructuredContent(), &pd)
-		if err != nil {
-			return nil, err
-		}
-		for _, ctn := range pd.Spec.Containers {
-			containers = append(containers, formatContainer(ctn, corev1.ContainerStatus{}, topController.GetCreationTimestamp()))
+
+		if workload.PodSpec != nil {
+			for _, ctn := range workload.PodSpec.Containers {
+				containers = append(containers, formatContainer(ctn, corev1.ContainerStatus{}, topController.GetCreationTimestamp()))
+			}
 		}
 		controller := formatControllers(topController.GetKind(), topController.GetName(), topController.GetNamespace(), string(topController.GetUID()), topController.GetOwnerReferences(), containers, topController.GetAnnotations(), topController.GetLabels())
-		controller.PodCount = podCount
+		controller.PodCount = float64(workload.RunningPodCount)
 		interfaces = append(interfaces, controller)
-
 	}
 
 	// Nodes