+++ plugins/ci/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 2.2.3
+* Update dependencies
+
+## 2.2.2
+* Update to go 1.19
+
 ## 2.2.1
 * Update versions