+++ plugins/falco-agent/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.2.2
+* Update dependencies
+
+## 0.2.1
+* Update to go 1.19
+
 ## 0.2.0
 * Build docker images for linux/arm64, and update to Go 1.19.1