+++ plugins/workloads/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 2.3.3
+* update controller-utils to 0.3.0
+## 2.3.2
+* Fix for top-level controllers with no pod info
+
+## 2.3.1
+* Update to go 1.19
+
 ## 2.3.0
 * Build docker images for linux/arm64, and update to Go 1.19.1